Fix 64 bit simulator warning.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@14957 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/font.c
@@ -306,7 +306,7 @@ struct font* font_load(const char *path)
     font_reset();
 
     /* currently, font loading replaces earlier font allocation*/
-    freeptr = (unsigned char *)(((int)mbuf + 3) & ~3);
+    freeptr = (unsigned char *)(((intptr_t)mbuf + 3) & ~3);
     fileptr = freeptr;
